Fix led_pattern_unlock and add led_pattern_lock
My observations indicated that led_pattern_unlock was tied to the
unlocks associated with unlocks.h.  However, the pattern used a hard
coded value for indicating unlocks which did not match the number of
unlocks in unlocks.h.  This change replaces the hard coded value with
MAX_ULCODES + 1 (for black badge) as indicated in unlocks.h.  It
additionally set's the last unlock (black badge) to be red instead of
green.

led_pattern_lock is a new function based off the
led_pattern_unlock_success pattern.  It is invoked when an unlock is
relocked by app-unlock.c.